Move Z3OpVisitor to its own file
